pyproject.toml
[project]
name = "google-drive-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.13.3",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
]
